+++ apps/mailbot/src/core/SlackClient.ts
