+++ apps/mailbot/src/env.ts
