+++ apps/mailbot/src/lambda.ts
