+++ apps/mailbot/src/scripts/auth-google.ts
@@ -1,4 +1,4 @@
-import { GoogleAuthorizer } from "../core/gmail/GoogleAuthorizer"
+import { GoogleAuthorizer } from "../core/gmail-client/GoogleAuthorizer"
 import { env } from "../env"
 
 const googleAuthorizer = new GoogleAuthorizer(env.GOOGLE_CLIENT_ID, env.GOOGLE_CLIENT_SECRET, env.GOOGLE_REDIRECT_URL)
