+++ apps/mailbot/src/scripts/labels.ts
