+++ apps/mailbot/src/utils.ts
