+++ apps/mailbot/tsup.config.ts
